Remove enums/constants pattern
