--- src/hutch_bunny/core/constants.py
@@ -1,7 +1,0 @@
-from hutch_bunny.core.enums import DistributionQueryType
-
-
-DISTRIBUTION_TYPE_FILE_NAMES_MAP = {
-    DistributionQueryType.DEMOGRAPHICS: "demographics.distribution",
-    DistributionQueryType.GENERIC: "code.distribution",
-}
