+++ src/hutch_bunny/core/solvers/query_solvers.py
@@ -4,14 +4,16 @@ from hutch_bunny.core.logger import logger
 
 from hutch_bunny.core.solvers.availability_solver import AvailabilitySolver
 from hutch_bunny.core.db_manager import SyncDBManager
-from hutch_bunny.core.rquest_models.query import AvailabilityQuery, DistributionQuery
+from hutch_bunny.core.rquest_models.query import (
+    AvailabilityQuery,
+    DistributionQuery,
+    DistributionQueryType,
+)
 from hutch_bunny.core.rquest_models.file import File
 from hutch_bunny.core.solvers.availability_solver import ResultModifier
 
 from hutch_bunny.core.rquest_models.result import RquestResult
-from hutch_bunny.core.enums import DistributionQueryType
 from hutch_bunny.core.settings import Settings
-from hutch_bunny.core.constants import DISTRIBUTION_TYPE_FILE_NAMES_MAP
 from hutch_bunny.core.solvers.demographics_solver import (
     DemographicsDistributionQuerySolver,
 )
@@ -103,7 +105,7 @@ def solve_distribution(
         result_file = File(
             data=res_b64,
             description="Result of code.distribution analysis",
-            name=DISTRIBUTION_TYPE_FILE_NAMES_MAP.get(query.code, ""),
+            name=query.code.file_name,
             sensitive=True,
             reference="",
             size=size,
